# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Integra AFC API
  version: 1.0.0
  description: Automated Fare Collection platform API. Versioned from v1; see docs/adr/
    for tenancy, auth and domain-model decisions.
paths:
...
  /api/v1/businesses/:
    get:
      operationId: businesses_list
      parameters:
      - name: limit
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: offset
        required: false
        in: query
        description: The initial index from which to return the results.
        schema:
          type: integer
      tags:
      - businesses
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedBusinessList'
          description: ''
    post:
      operationId: businesses_create
      tags:
      - businesses
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Business'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Business'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Business'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Business'
          description: ''
...
components:
  schemas:
...
    BookingModeDefaultEnum:
      enum:
      - reservation
      - tap_and_go
      type: string
      description: |-
        * `reservation` - Reservation
        * `tap_and_go` - Tap and go
...
    Business:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        vertical:
          $ref: '#/components/schemas/VerticalEnum'
        name:
          type: string
          maxLength: 255
        currency:
          type: string
          maxLength: 8
        timezone:
          type: string
          maxLength: 64
        booking_mode_default:
          $ref: '#/components/schemas/BookingModeDefaultEnum'
        fare_pricing_mode:
          $ref: '#/components/schemas/FarePricingModeEnum'
        is_active:
          type: boolean
        kyb_status:
          allOf:
          - $ref: '#/components/schemas/KybStatusEnum'
          readOnly: true
        kyb_submitted_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - booking_mode_default
      - created_at
      - currency
      - id
      - kyb_status
      - kyb_submitted_at
      - name
      - timezone
      - vertical
...
    FarePricingModeEnum:
      enum:
      - flat
      - per_segment
      type: string
      description: |-
        * `flat` - Flat
        * `per_segment` - Per segment
...
    KybStatusEnum:
      enum:
      - pending
      - submitted
      - approved
      - rejected
      type: string
      description: |-
        * `pending` - Pending
        * `submitted` - Submitted
        * `approved` - Approved
        * `rejected` - Rejected
...
    PaginatedBusinessList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=400&limit=100
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=200&limit=100
        results:
          type: array
          items:
            $ref: '#/components/schemas/Business'
...
    VerticalEnum:
      enum:
      - shuttle
      - intercity
      - metro
      type: string
      description: |-
        * `shuttle` - Shuttle
        * `intercity` - Intercity
        * `metro` - Metro
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT